requirements.txt
-e git+https://github.com/facebookresearch/fastText.git#egg=elasticutils
pybind11
tqdm
nltk
numpy
torchvision_nightly
torch_nightly
fastai
